[defs] Add docstrings to most functions and methods
The docstrings are generated from the gobject-introspection data by a
very simple no-brainer script and describe the C-API only. They are
neither beautified nor adapted to match the Python API.

This is far from perfect, but at least it's an entry point for possible
volunteers (please?) to improve our horrible documentation.