Use an explicitly sized string for Stage.read_pixels() return value
The return value for clutter.Stage.read_pixels() is a binary blob
representing a linear buffer of RGBA 8bit data with a rowstride of
4 * width. Using Py_BuildValue("s") is not a good idea, as it assumes
that the string is going to be NUL-terminated. We should use the "s#"
signature and pass the length of the buffer instead.

Signed-off-by: Emmanuele Bassi <ebassi@linux.intel.com>
(cherry picked from commit c05436ed9c236c060628e5a09ce8d32ade58cca2)