Rework how types and symbols are handled
Instead of clumping everything into a single file and updating the
minimum required version of Clutter, we should do what pygtk does: move
types and symbols into version-specific defs files and add them to the
build depending on the version of Clutter we're compiling against.

This massive commit it's the minimum amount of work needed to make this
mess actually compile.

First, we need to use our own copy of the h2def.py; apparently, the
upstream version blatantly *lies* about the -f|--defsfilter argument: it
says that you can add different defs file to incrementally add symbols
for a specific version, but in reality only the last one will be
considered. EPIC. FAIL. The fact that upstream never noticed this makes
me think that they write the defs file by hand, which is epically lame
and fully defeats the point of auto-generation. I had my doubts, but
this last fiasco makes me long for the day when the utter brain damage
that is pygtk will be deprecated and purged from git.gnome.org in favour
of the sane approach of the introspection-based bindings.

Then we need to clean up the build environment to take into account
version-specific files; we also take the chance to depend on Automake
≥ 1.11 and drop Shave for good.

Finally, we need to remove all the 1.2 symbols from the base defs file
and move them into the clutter-1.2 definitions.

Again, this commit is the minimal amount of work needed to make this
thing *compile*. There are a ton and a half of fixes required to make it
work.